requirements.txt
langchain
langchain-community
langchain_ollama
faiss-cpu
pypdf
flask
werkzeug
pydantic
flask-cors